Disallow multiple submissions for an Exam/Game qInstance.
This required refactoring all *TestServer.js files to be CommonJS
modules instead of AMD modules. In the future, we should limit AMD
modules to frontend code only, as AMD modules can only require other
frontend modules and not backend modules, which is severely limiting for
backend code.

This also modifies `updateWithSubmission()` to be asynchronous.

Closes #307.