Render each submission into a separate panel
To be consistent with the behavior of question.html, answer.html, each
submission now gets it's own panel and renders a copy of submission.html
within it. The submission meta-data (index, date,
correct/incorrect/saved) are incorporated into the panel heading.

I would like to put the blue/green/red labels used elsewhere in the
panel heading (instead of just the words), but they don't seem to draw
properly. Instead, I've set the panel style to change based on the
submission status (info/success/danger) once we have access to it.