Import require-frontend after reading config
This was a regression caused by e330482 ("split json-load.js and
course-db.js out of server.js"). course-db requires require-frontend,
which sets up the config for requirejs. However, when course-db is
required, config.json hasn't been read yet. As a result, requirejs is
loaded with the default config, which has 'courseDir:
"../exampleCourse"'. This, in turn, breaks course code that uses
"serverCode" and "clientCode".