Working prototype of MATLAB I/O elements and example question
Students are asked to multiply two matrices. Submitted answer is checked for correctness with respect to relative and absolute tolerance, number of significant digits, or number of digits after the decimal. TODOs remain, some because PL elements are not fully implemented yet. Wanted to get the prototype up for initial comment.