explicitly unlink broken variants; add assessment error
The previous approach created unlinked variants (that is, a variant
that has instance_question_id = NULL). The problem with this is that
the resulting error rows were then unable to determine the corresponding
question and course information for the variant, because they do it by
following instance_question_id.

This commit changes to an approach where we create the variant with
the instance_question_id linked, then we insert the error rows (which
can find the question/course), and then we unlink the broken variants by
setting instance_question_id to NULL.