Create correct links on courseQuestions assessment badges
These now link into the courseInstance instructor page; permissions are
checked when the user follows the link.

Right now, if a user does NOT have instructor permissions on that
course_instance, it downgrades them to student permissions and creates a
new assessment instance (or takes you to the quiz start page). Is this
the desired behavior?