Add a 'courseGradingJob' page and routing
Right now the only access to 'instructorGradingJob' is through the 'instructorAssessment' page. There is no analogue at the course level, so this page doesn't do anything right now.

Since 'instructorQuestion' and 'courseQuestion' are the same page now, adding access to grading jobs from this point would require the course-level routing in server.js

It's quite possible that this page and 'instructorGradingJob' can be combined into a single access point when we make the change mentioned above.