Fix "Edit on GitHub" linkin mkdocs.yml
Trying to fix the "Edit on GitHub" link displayed on ReadTheDocs pages. We are using doc/ instead of docs/ which ruins the default setting if this isn't specified, and also I have directed it to `blob` instead of `edit` to arrive at a viewing page rather than an editing page.