Fix in pl-code.py due to hljs bug
hljs has a compatibility issue where a single-line C++ comment can't be parsed correctly if the line is "highlighted" by pl-code, and then the line following the single-line comment is also marked as a comment by hljs. Inserting a terminating \n on the highlighted line, inside the span, seems to fix this.