Fix external grading results containing NULL bytes
This is an update to #935. That PR failed to fix all cases of NULL
bytes, because NULL bytes might be encoded as "\u0000" in JSON. To fix
this, we process the parsed JSON as a JavaScript object, recursively
searching for strings and replacing NULL bytes in them. We also change
from #935 in that we don't just remove NULL bytes, but rather replace
them with the Unicode replacement character "\uFFFD".

Closes #1270