Fix locals for externally-graded question panels
When we render the new question panels for externally graded results, we
have been triggered by the grading results queue and so we are not
inside a request. This means we don't have res.locals, but rendering
assumes we have it, so we construct a fake locals object that has all
the same info. We had missed locals.devMode, which was subtly changing
the caching behavior in dev. To fix this, and hopefully ensure we keep
the real locals and fake locals the same in the future, this PR adds a
helper function to create the base locals object, and makes both normal
requests and the fake locals creator use the helper.