Fix tryLock() to end the transaction if the lock was not acquired.
It turns out that #2395 did not completely fix the tryLock() handling. If we fail to acquire the lock, we need to both return null _and_ end the transaction. The change in #2395 fixed the "return null" part, but it turns out that b190cdc207e248cb6229d32fde5836d3374cac66 had also eliminated the code that ended the transaction if the lock was null. This PR add this back.

I also added more comments to explain how the locking logic works.