Remove superfluous "* 2" on color palette
The palette shown already specifies all 16 colors explicitly, so there's no need to duplicate to get the second 8 (the "bright" set). Coincidentally, the library automatically generates a "bright" variant of the first eight colors for the second eight, if they aren't already specified.

By the way, the built-in "xterm" palette is very similar to the one defined here.