Replace `correct_answer` with `correct_answers`
I had a look at all of the example problems and all of them use `data["correct_answers"]` rather than `data["correct_answer"]`. In some cases in the docs, `correct_answer` is used and generates an error (specifically in the symbolic_input question.

It would be nice if someone could confirm this is the case before updating the documentation.