Fix docs to update `correct_answer` to `correct_answers` (#4194)
* Replace `correct_answer` with `correct_answers`

I had a look at all of the example problems and all of them use `data["correct_answers"]` rather than `data["correct_answer"]`. In some cases in the docs, `correct_answer` is used and generates an error (specifically in the symbolic_input question.

It would be nice if someone could confirm this is the case before updating the documentation.

* Reverted the HTML changes due to my misunderstanding

h/t @jonatanschroeder

* change hyphen to underscore 

`correct_answers` instead of `correct-answers`

* Change python from correct-answers to correct_answers

I left the HTML as `correct-answer` for now because I'm not sure if that's wrong (hyphen vs. dash).

* Update comment with correct_answerS