A few changes to the apt-get stuff. (See expanded commit message below.)
Changed "apt" to "apt-get" which is more appropriate for non-interactive usage.
Adding gedit and less, which some students may expect to be there.
Adding the manpages back for the same reason. (Putting this early makes the
build faster than putting it later, too.)
Fixing a mispelling on the defaults directory.