Fix error handling in pl-matching if params is not set (#4830)
I had problems with this for a student that opened a question that had been released by mistake before it was properly set, but this can also be a problem if a question is changed after a submission had been made (bad practice, but should have a cleaner solution). Current version creates issues every time the question is opened. This should remove that problem