Create pl-truth-table element
Create a truth table element with support for custom true/false labels, custom rows, inline support of pl elements for variable labels, custom rows and numbers of rows, and user defined expressions. An instance of the element can be found in both the testCourse/elements directory and the top-level elements directory. A training question can be found in the testCourse/questions directory with the name truthTableDemo. The README of this question also has more detailed notes about the element's features and use.