Try adding "modules" id to the existing heading
This is hacky, but putting `<a id="..."></a>` on the heading adds another invisible ID that can be linked. It's important not to put a space before the `<a` or else the implicitly-generated anchor may get an extra `-` appended at the end, e.g. `assessment-modules` -> `assessment-modules-`. My goal here is to allow _either_ `assessment-modules` or `modules` to tag the same heading.