use Ubuntu 24.04, Miniforge like Python grader, mamba instead of conda
In addition to using Miniforge, we activate mamba along with conda
(based on the Miniforge installation recommendations). Also, allowing
mamba to install pygraphviz separately first avoids the compilation
issues with pygraphviz. There may be other benefits to using mamba
over conda: it doesn't seem to have as many dynamic linking errors,
and it seems to be faster.

The Ubuntu Noble (24.04) base has newer gcc than the plain ubuntu
base (which seems to be Focal). That would be important if we need
to install gcc later on with parity compared to Amazon Linux 2023
like the Python grader image uses. However, installing pygraphviz
with mamba currently works without compilation, at least on x86_64.